Taxi_data_de_code/Taxi_data_cloud_function_auto_piepeline_requirement.txt
functions-framework==3.*
google-cloud-storage
google-cloud-bigquery
pandas
fastparquet
pyarrow